requirements.txt
beautifulsoup4==4.11.1
google_auth_oauthlib==0.7.1
lxml==4.9.2
matplotlib==3.6.2
moviepy==1.0.3
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
protobuf==4.21.11
pytesseract==0.3.10
pillar-youtube-upload==0.3.0
requests==2.28.1
selenium==3.8.0
textdistance==4.5.0
opencv-python==4.6.0.66
openpyxl==3.0.10
webdriver-manager==3.8.5
